requirements.txt
pyTelegramBotAPI==3.8.2
python-dotenv==0.19.0
requests==2.26.0
libretranslatepy==2.1.1